fix(api): use new uploads API
This change is using the new serverless-upload API instead of the AWS S3 upload. This drops also the
support for the AWS Upload function

BREAKING CHANGE: Removes uploadToAws as a function
