+++ src/utils/content-type.ts
@@ -1,0 +1,28 @@
+/** @module @twilio-labs/serverless-api/dist/utils */
+
+import fileType from 'file-type';
+import mime from 'mime-types';
+
+/**
+ * Tries to determine the content type of a string or buffer
+ *
+ * @export
+ * @param {(string | Buffer)} content the content to check
+ * @returns {(string | undefined)} a valid content type or undefined
+ */
+export function getContentType(content: string | Buffer): string | undefined {
+  let contentType: string | undefined;
+
+  if (typeof content !== 'string') {
+    const type = fileType(content);
+    if (type) {
+      contentType = type.mime;
+    }
+  }
+
+  if (name && !contentType) {
+    contentType = mime.contentType(name) || undefined;
+  }
+
+  return contentType;
+}
