+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 /** @module @twilio-labs/serverless-api/dist/utils */
 
-export * from './aws-upload';
+export * from './content-type';
 export * from './fs';
 export * from './sleep';
